Code refactoring
-Added hooks
-Code cleanup
