+++ app/screens/ListingEditScreen.js
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { StyleSheet } from "react-native";
 import * as Yup from "yup";
-import * as Location from "expo-location";
 import {
   AppForm,
   AppFormField,
@@ -11,6 +10,7 @@ import {
 import Screen from "../components/Screen";
 import FormImagePicker from "../components/forms/FormImagePicker";
 import CategoryPickerItem from "../components/CategoryPickerItem";
+import useLocation from "../hooks/useLocation";
 
 const validationSchema = Yup.object().shape({
   title: Yup.string().required().min(1).label("Title"),
@@ -31,18 +31,7 @@ const categories = [
 ];
 
 function ListingEditScreen() {
-  const [location, setLocation] = useState();
-  const getLocation = async () => {
-    const { granted } = await Location.requestBackgroundPermissionsAsync();
-    if (!granted) return;
-    const {
-      coords: { latitude, longitude },
-    } = await Location.getLastKnownPositionAsync();
-    setLocation({ latitude, longitude });
-  };
-  useEffect(() => {
-    getLocation();
-  }, []);
+  const location = useLocation();
   return (
     <Screen style={styles.container}>
       <AppForm
